requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Google Cloud Services
google-cloud-storage>=2.18.0
google-cloud-pubsub>=2.23.0

# Google Generative AI
google-genai>=1.47.0

# Environment Variables
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-cov>=7.0.0

# Utilities
numpy>=1.26.0
psutil>=5.9.0